niri_ipc
Enum Event

pub enum Event {
    WorkspacesChanged {
        workspaces: Vec<Workspace>,
    },
    WorkspaceActivated {
        id: u64,
        focused: bool,
    },
    WorkspaceActiveWindowChanged {
        workspace_id: u64,
        active_window_id: Option<u64>,
    },
    WindowsChanged {
        windows: Vec<Window>,
    },
    WindowOpenedOrChanged {
        window: Window,
    },
    WindowClosed {
        id: u64,
    },
    WindowFocusChanged {
        id: Option<u64>,
    },
    KeyboardLayoutsChanged {
        keyboard_layouts: KeyboardLayouts,
    },
    KeyboardLayoutSwitched {
        idx: u8,
    },
}

A compositor event.

Variants

WorkspacesChanged

The workspace configuration has changed.

Fields

workspaces: Vec<Workspace>

The new workspace configuration.

This configuration completely replaces the previous configuration. I.e. if any workspaces are missing from here, then they were deleted.

WorkspaceActivated

A workspace was activated on an output.

This doesnโt always mean the workspace became focused, just that itโs now the active workspace on its output. All other workspaces on the same output become inactive.

Fields

id: u64

Id of the newly active workspace.

focused: bool

Whether this workspace also became focused.

If true, this is now the single focused workspace. All other workspaces are no longer focused, but they may remain active on their respective outputs.

WorkspaceActiveWindowChanged

An active window changed on a workspace.

Fields

workspace_id: u64

Id of the workspace on which the active window changed.

active_window_id: Option<u64>

Id of the new active window, if any.

WindowsChanged

The window configuration has changed.

Fields

windows: Vec<Window>

The new window configuration.

This configuration completely replaces the previous configuration. I.e. if any windows are missing from here, then they were closed.

WindowOpenedOrChanged

A new toplevel window was opened, or an existing toplevel window changed.

Fields

window: Window

The new or updated window.

If the window is focused, all other windows are no longer focused.

WindowClosed

A toplevel window was closed.

Fields

id: u64

Id of the removed window.

WindowFocusChanged

Window focus changed.

All other windows are no longer focused.

Fields

id: Option<u64>

Id of the newly focused window, or None if no window is now focused.

KeyboardLayoutsChanged

The configured keyboard layouts have changed.

Fields

keyboard_layouts: KeyboardLayouts

The new keyboard layout configuration.

KeyboardLayoutSwitched

The keyboard layout switched.

Fields

idx: u8

Index of the newly active layout.