niri_ipc

Struct Output

pub struct Output {
    pub name: String,
    pub make: String,
    pub model: String,
    pub serial: Option<String>,
    pub physical_size: Option<(u32, u32)>,
    pub modes: Vec<Mode>,
    pub current_mode: Option<usize>,
    pub vrr_supported: bool,
    pub vrr_enabled: bool,
    pub logical: Option<LogicalOutput>,
}

Connected output.

Fields
name: String

Name of the output.

make: String

Textual description of the manufacturer.

model: String

Textual description of the model.

serial: Option<String>

Serial of the output, if known.

physical_size: Option<(u32, u32)>

Physical width and height of the output in millimeters, if known.

modes: Vec<Mode>

Available modes for the output.

current_mode: Option<usize>

Index of the current mode in Self::modes.

None if the output is disabled.

vrr_supported: bool

Whether the output supports variable refresh rate.

vrr_enabled: bool

Whether variable refresh rate is enabled on the output.

logical: Option<LogicalOutput>

Logical output information.

None if the output is not mapped to any logical output (for example, if it is disabled).