niri_ipc
Module state

Helpers for keeping track of the event stream state.

Create an EventStreamState using Default::default(), or any individual state part if you only care about part of the state.

Connect to the niri socket and request an event stream.

Pass every Event to EventStreamStatePart::apply on your state.

Read the fields of the state as needed.

Structs

EventStreamState

The full state communicated over the event stream.

KeyboardLayoutsState

The keyboard layout state communicated over the event stream.

WindowsState

The windows state communicated over the event stream.

WorkspacesState

The workspaces state communicated over the event stream.

Traits

EventStreamStatePart

Part of the state communicated via the event stream.